requirements.txt
aiohttp==3.7.4.post0
apturl==0.5.2
asn1crypto==0.24.0
async-timeout==3.0.1
attrs==21.2.0
Brlapi==0.6.6
certifi==2021.5.30
chardet==4.0.0
charset-normalizer==2.0.6
command-not-found==0.3
cryptography==2.1.4
cupshelpers==1.0
defer==1.0.6
discord.py==1.7.3
distro-info===0.18ubuntu0.18.04.1
httplib2==0.19.1
idna==3.2
idna-ssl==1.1.0
keyring==10.6.0
keyrings.alt==3.0
language-selector==0.1
launchpadlib==1.10.6
lazr.restfulclient==0.13.5
lazr.uri==1.0.3
louis==3.5.0
macaroonbakery==1.1.3
Mako==1.0.7
MarkupSafe==1.0
multidict==5.1.0
netifaces==0.10.4
oauth==1.0.1
oauth2==1.9.0.post1
olefile==0.45.1
pexpect==4.2.1
Pillow==5.1.0
protobuf==3.0.0
psutil==5.4.2
pycairo==1.16.2
pycrypto==2.6.1
pycups==1.9.73
pygobject==3.26.1
pymacaroons==0.13.0
PyNaCl==1.1.2
pyparsing==2.4.7
pyRFC3339==1.0
python-apt==1.6.5+ubuntu0.3
python-dateutil==2.6.1
python-debian==0.1.32
python-dotenv==0.19.0
pytz==2018.3
pyxdg==0.25
PyYAML==3.12
reportlab==3.4.0
requests==2.26.0
requests-unixsocket==0.1.5
SecretStorage==2.3.1
simplejson==3.13.2
six==1.11.0
system-service==0.3
systemd-python==234
typing-extensions==3.10.0.2
ubuntu-drivers-common==0.0.0
ufw==0.36
unattended-upgrades==0.1
urllib3==1.26.7
usb-creator==0.3.3
vboxapi==1.0
wadllib==1.3.2
xkit==0.0.0
yarl==1.6.3
zope.interface==4.3.2
